requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
pillow==12.1.1
python-decouple==3.8
python-telegram-bot==20.7
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
